# Conflicts:
#	codeflash/api/aiservice.py
#	codeflash/languages/base.py
#	codeflash/languages/java/support.py
#	codeflash/languages/javascript/support.py
#	codeflash/languages/python/support.py
#	codeflash/verification/verifier.py

+++ tests/test_languages/test_java/test_instrumentation.py
@@ -219,6 +219,7 @@ public class FibonacciTest {
     }
 }
 """
+
         test_file.write_text(source)
 
         func = FunctionToOptimize(
@@ -2688,7 +2689,7 @@ public class CounterTest__perfinstrumented {
                 }
             }
         }
-        assertEquals(1, _cf_result1_1);
+        assertEquals(1, (int)_cf_result1_1);
     }
 }
 """
